Convert to long in order to support more ambitious test cases.

+++ DynamicProgramming.UnitTests/FibonacciBaseTests.cs
@@ -19,7 +19,7 @@ namespace DynamicProgramming.UnitTests
             Assert.Equal(0, BaseCase(0));
         }
 
-        public override int Fib(int n)
+        public override long Fib(long n)
         {
             throw new System.NotImplementedException();
         }
